verify user has access to player being deleted; add tests

+++ dashboard-server/src/main/java/com/footballstatsdashboard/services/PlayerService.java
@@ -34,6 +34,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.OptionalDouble;
 import java.util.UUID;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 import static com.footballstatsdashboard.core.utils.Constants.PLAYER_ATTRIBUTE_CATEGORY_MAP;
@@ -52,14 +53,7 @@ public class PlayerService {
 
     public Player getPlayer(UUID playerId) {
         ResourceKey resourceKey = new ResourceKey(playerId);
-
-        try {
-            return this.playerDAO.getDocument(resourceKey, Player.class);
-        } catch (DocumentNotFoundException documentNotFoundException) {
-            String errorMessage = String.format("No player entity found for ID: %s", playerId);
-            LOGGER.error(errorMessage);
-            throw new ServiceException(HttpStatus.NOT_FOUND_404, errorMessage);
-        }
+        return fetchPlayerData(playerId, resourceKey);
     }
 
     public Player createPlayer(Player incomingPlayer, Club clubDataForNewPlayer, String createdBy) throws IOException {
@@ -189,9 +183,18 @@ public class PlayerService {
         return updatedPlayer;
     }
 
-    // TODO: 1/3/2022 add checks to make sure player being deleted belongs to the club and user making the request
-    public void deletePlayer(UUID playerId) {
+    public void deletePlayer(UUID playerId, Predicate<UUID> doesClubBelongToUser) {
         ResourceKey resourceKey = new ResourceKey(playerId);
+
+        // verify that the current user has access to the player they are trying to delete
+        // since a player cannot belong to more than one club, we can transitively check the user's access to the player
+        // by checking their access to the club the user belongs to
+        Player player = fetchPlayerData(playerId, resourceKey);
+        if (!doesClubBelongToUser.test(player.getClubId())) {
+            LOGGER.error("Player with ID: {} does not belong to user making request", player.getId());
+            throw new ServiceException(HttpStatus.FORBIDDEN_403, "User does not have access to this player!");
+        }
+
         try {
             this.playerDAO.deleteDocument(resourceKey);
         } catch (DocumentNotFoundException documentNotFoundException) {
@@ -251,6 +254,16 @@ public class PlayerService {
                     playerAttributes);
             throw new ServiceException(HttpStatus.INTERNAL_SERVER_ERROR_500,
                     "Something went wrong trying to calculate current ability of player");
+        }
+    }
+
+    private Player fetchPlayerData(UUID playerId, ResourceKey resourceKey) {
+        try {
+            return this.playerDAO.getDocument(resourceKey, Player.class);
+        } catch (DocumentNotFoundException documentNotFoundException) {
+            String errorMessage = String.format("No player entity found for ID: %s", playerId);
+            LOGGER.error(errorMessage);
+            throw new ServiceException(HttpStatus.NOT_FOUND_404, errorMessage);
         }
     }
 }
